No compiler warnings + Returns null on no key
Really happy that I got SeaJson (and Buttermilk in the future 0.0.3 update) to have no compiler warnings :) Also now returns null if the key passed to the library is not in the dict passed in. No longer returns "NO INPUT" on no input as it will now return null due to the previous addition I made.